fastapi/requirements.txt
uvicorn
fastapi
pymongo
python-dotenv